ServerLaunchEvent now supports asynchronous loading tasks from Kotlin. The server will not go online until these tasks complete, making them useful for important blocking loading operations that need to be ran from scripts.
There is also now a ServerShutdownEvent which is posted when the server is gracefully shutting down. Similarly to ServerLaunchEvent, asynchronous tasks can be ran through this and the server will not shutdown until they are complete.

Lazy-posting is the final feature of this important update, which allows for more control how events are applied. It allows you to delay the actual posting of the events until a later time