feat: rework polygon area
The old implementation relied on the java awt polygon implementation. In this use case, the backing polygon is immutable and as a result, some optimizations can be done such as precomputing the bounding box/constraints.

The previous contains lookup creates new rectangles with every contains call and then did some math due to rect storing the width/height. In this use case the width/height are not needed so we can just store the upper and lower bounds.

Future work would be tweaking the input from using a List of Points to something that used Ints explicitly for the position. Additionally, the way the hashcode is generated is suboptimal for this use case.

For the equals method, it could be set to just compare the hashcode if both are of a similar type of PolygonArea.

Since every area stores positions/position set, there is also the option of bitpacking all positions into ints or just checking the bounding box and then calling contains on that set for the desired position